perf: 去除多余的文本解析为具体类型的过程

+++ VModer.Core/Models/BuildingInfo.cs
@@ -1,7 +1,7 @@
 ﻿namespace VModer.Core.Models;
 
-public sealed class BuildingInfo(string name, ushort? maxLevel)
+public sealed class BuildingInfo(string name, int? maxLevel)
 {
 	public string Name { get; } = name;
-	public ushort? MaxLevel { get; } = maxLevel;
+	public int? MaxLevel { get; } = maxLevel;
 }
